Factorize infill_criterion optmization between TREGO and EGO
